Remove usage of generic wildcard type by refactoring UnixTimeFormatter to interface
Also some encapsulation improvements.
